add tag & editor page & gallery page
